Back out 02_changeset_r52_netwm_full_screen.patch from Debian (closes #22)
The only thing the patch seems to do is try to calculate the window borders so
it can move the window to the right place. However, XResizeWindow simply does
not move the window no matter how wide the borders are, which is the intended
behaviour. Plus, certain window managers (like fluxbox) report an incorrect
window border width.

I have yet to test this commit on a Xinerama system, though. It _might_ affect
something there.