slideshow.c: save_image: Fix handling of capital file endings (closes #51)
gib_imlib_save_image_with_error_return was used, but it tries to set the Imlib
image format by examining the filename to save to.  However, it only supports
lowercase file endings (as in .jpg) and breaks with others (like .JPG).

Since giblib is unmaintained, for now the fix is only used here.