Show correct file number after deleting image
In the image_remove function, slideshow_change_image needs to be called before
removing the image from the filelist.  Because of this, the "x of y" displayed
by --draw-filename is wrong (y is 1 too high).

This commit introduces a new 'render' argument to slideshow_chingae_image.
The image_remove function calls it with render=0, edits the filelist and then
runs by winwidget_render_image itself.

Patch by Yu-Jie Lin.