Add --draw-tinted option to draw overlay text on a semi-transparent background (closes #60, #61)
By default, feh draws white text with a black drop-shadow, which is well
readable in most, but not all cases.  The --draw-tinted option draws a
semi-transparent black box around all texts.  This improves readability, but
does not look as nice, so it won't become default behaviour.

It was suggested to use #7f7f7f7f for the background, but I decided that
for small images surrounded by a black border.