feh_reload_image: Always die if reload failed
This fixes a segfault in --reload when the image becomes unloadable while feh
is running.

Note that this does not affect --reload with more than one file: If the
current image is removed, feh will simply jump to the next image. feh will
only quit if the file still exists, but can no longer be loaded.