feh_reload_image: Try to reload unloadable (still existing) images ad infinitum
This commit removes the feh_file_remove_from_list call from feh_reload_image.
We can safely do this because if the file itself was removed, cb_reload_timer
will take care of updating the filelist, and in all other cases it is intended
to leave the image in the filelist and retry.

A warning will be displayed in the image if the reload failed.