Add --image-bg option for wallpaper border
Previously feh would set a black background around wallpaper images,
with no option to change.

This commit extends the functionality offered by the --image-bg option,
which is available for regular image viewing, to centered and max'ed feh
backgrounds. Essentially, this means a white border can now be chosen
in addition to a black one.