Fix dead code
The check if buffer == NULL is always false, since buffer is an
autoamtic variable allocated when entering the function.  What we
instead want to do is to check if the string is empty after the call to
exif_get_info(), since that means we could not read any exif
information.
When the code once more is enabled, I discovered that we need to copy
the information string into info_buf as well as into  buffer, since it
is the former that is used to print the exif information on top of the
picture.  Without this change, imlib warns about trying to write NULL
strings.